pyproject.toml
[project]
name = "texhelper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "numpy>=2.1.3",
    "onnx>=1.17.0",
    "onnxruntime>=1.20.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.17",
    "transformers>=4.46.2",
    "uvicorn>=0.32.0",
]
